[script] Add initial version for the bundler script
Bundler script is a naive reference implementation for the set of
operations required to use the "ld -r" relocatable object alongside
objcopy when creating a visibility-focused static bundle object.

Format supports both a regex for kept global symbols, as well as
a version script syntax as is usually used for shared objects.

Support for stripping the symbols of localized symbols will require
ld support, and is more than just a "tooling" issue.

Signed-off-by: Eyal Itkin <eyal.itkin@gmail.com>